requirements.txt
appdirs==1.4.3
atomicwrites==1.3.0
attrs==19.3.0
backcall==0.1.0
black==19.10b0
bleach==3.1.1
cachetools==4.0.0
certifi==2019.11.28
cfgv==3.1.0
chardet==3.0.4
Click==7.0
colorama==0.4.3
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
distlib==0.3.0
entrypoints==0.3
filelock==3.0.12
fuzzywuzzy==0.18.0
google-api-python-client==1.7.11
google-auth==1.11.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
httplib2==0.17.0
identify==1.4.11
idna==2.9
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.16.0
Jinja2==2.11.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.0.0
jupyter-console==6.1.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.6.3
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
kiwisolver==1.1.0
lxml==4.5.0
MarkupSafe==1.1.1
matplotlib==3.1.3
mistune==0.8.4
more-itertools==8.2.0
nbconvert==5.6.1
nbformat==5.0.4
nodeenv==1.3.5
notebook==6.0.3
numpy==1.18.1
oauthlib==3.1.0
packaging==20.3
pandas==1.0.1
pandocfilters==1.4.2
parso==0.6.1
pathspec==0.7.0
patsy==0.5.1
pickleshare==0.7.5
pluggy==0.13.1
pre-commit==2.1.1
prometheus-client==0.7.1
prompt-toolkit==3.0.3
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycurl==7.43.0.5
Pygments==2.5.2
pyparsing==2.4.6
pyrsistent==0.15.7
pytest==5.4.1
python-box==4.2.1
python-dateutil==2.8.1
python-dotenv==0.12.0
python-Levenshtein==0.12.0
pytz==2019.3
pywin32==227
pywinpty==0.5.7
PyYAML==5.3
pyzmq==19.0.0
qtconsole==4.7.1
QtPy==1.9.0
regex==2020.2.20
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
scipy==1.4.1
Send2Trash==1.5.0
six==1.14.0
statsmodels==0.11.1
terminado==0.8.3
testpath==0.4.4
toml==0.10.0
tornado==6.0.3
traitlets==4.3.3
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.8
virtualenv==20.0.7
wcwidth==0.1.8
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.0.0
